alura-django-rest-school/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
markdown==3.4.1
django-cors-headers==3.13.0
validate-docbr==1.10.0
Pillow==9.2.0
Faker==14.1.0
django-admin-honeypot-updated-2021==1.2.0
isort
black
